+++ lib/td_dq/search/mappings.ex
@@ -4,9 +4,10 @@ defmodule TdDq.Search.Mappings do
   """
   alias TdCache.TemplateCache
 
+  @raw_sort %{raw: %{type: "keyword"}, sort: %{type: "keyword", normalizer: "sortable"}}
+
   def get_mappings do
     content_mappings = %{properties: get_dynamic_mappings("dq")}
-    content_mappings_bg = %{properties: get_dynamic_mappings("bg")}
 
     mapping_type = %{
       id: %{type: "long"},
@@ -36,16 +37,15 @@ defmodule TdDq.Search.Mappings do
         properties: %{
           id: %{type: "long"},
           name: %{type: "text", fields: %{raw: %{type: "keyword", normalizer: "sortable"}}},
-          content: content_mappings_bg
+          content: %{
+            properties: get_dynamic_mappings("bg", "user")
+          }
         }
       },
       updated_at: %{type: "date", format: "strict_date_optional_time||epoch_millis"},
       inserted_at: %{type: "date", format: "strict_date_optional_time||epoch_millis"},
       goal: %{type: "long"},
       minimum: %{type: "long"},
-      weight: %{type: "long"},
-      population: %{type: "text", fields: %{raw: %{type: "keyword"}}},
-      priority: %{type: "text", fields: %{raw: %{type: "keyword"}}},
       df_name: %{type: "text", fields: %{raw: %{type: "keyword"}}},
       rule_type: %{
         properties: %{
@@ -80,20 +80,20 @@ defmodule TdDq.Search.Mappings do
     %{mappings: %{_doc: %{properties: mapping_type}}, settings: settings}
   end
 
-  defp get_dynamic_mappings(scope) do
+  def get_dynamic_mappings(scope, type \\ nil) do
     scope
     |> TemplateCache.list_by_scope!()
-    |> Enum.flat_map(&get_mappings(&1, scope))
+    |> Enum.flat_map(&get_mappings(&1, scope, type))
     |> Enum.into(%{})
   end
 
-  defp get_mappings(%{content: content}, "bg") do
+  defp get_mappings(%{content: content}, "bg", type) do
     content
-    |> Enum.filter(&(Map.get(&1, "type") == "user"))
+    |> Enum.filter(&(Map.get(&1, "type") == type))
     |> Enum.map(&field_mapping/1)
   end
 
-  defp get_mappings(%{content: content}, _scope) do
+  defp get_mappings(%{content: content}, _scope, _type) do
     Enum.map(content, &field_mapping/1)
   end
 
@@ -107,6 +107,10 @@ defmodule TdDq.Search.Mappings do
 
   defp field_mapping(%{"name" => name, "type" => "enriched_text"}) do
     {name, mapping_type("enriched_text")}
+  end
+
+  defp field_mapping(%{"name" => name, "type" => "user"}) do
+    {name, %{type: "text", fields: @raw_sort}}
   end
 
   defp field_mapping(%{"name" => name, "values" => values}) do