+++ test/td_dq/rules/aggregations_test.exs
@@ -4,7 +4,10 @@ defmodule TdDq.Search.AggregationsTest do
   alias TdDq.Search.Aggregations
 
   def create_template(template) do
-    TemplateCache.put(template)
+    template
+    |> Map.put(:updated_at, DateTime.utc_now())
+    |> TemplateCache.put()
+
     template
   end